tests/wpt/tests/css/css-ui/parsing/outline-width-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS UI Level 4: parsing outline-width with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-ui/#outline-width">
<meta name="assert" content="outline-width supports the full grammar '<border-width>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("outline-width", "0", "0px");
test_valid_value("outline-width", "1px");
test_valid_value("outline-width", "2em");
test_valid_value("outline-width", "calc(2em + 3ex)");
test_valid_value("outline-width", "thin");
test_valid_value("outline-width", "medium");
test_valid_value("outline-width", "thick");
</script>
</body>
</html>
